requirements.txt
Flask
python-dotenv
pymongo
authlib
azure-ai-inference
azure-core
